+++ provider/component/all.go
@@ -1,9 +1,11 @@
-// Package component provides components for the NetBird Pulumi provider.
 package component
 
 import "github.com/pulumi/pulumi-go-provider/infer"
 
 // All returns all registered provider components.
 func All() []infer.InferredComponent {
-	return []infer.InferredComponent{}
+	return []infer.InferredComponent{
+		infer.Component(&NetworkBundle{}),
+		infer.Component(&DNSZoneBundle{}),
+	}
 }